Fix drawing of urgency border colour
It looks as though commit f37c14902fd639b87090659433d815f20fe3a42a
introduced a regression in the way urgency borders are rendered.  Because
each window is considered across all desktops for monitors, we must now
augment the checking of an urgent client when looking up the border colours
for focused windows also.

Without this, a single window on a desk on an unfocused monitor will be
correctly set as having an urgent hint, but its borders will not pick up on
any urgent border colour settings.